requirements.txt
audeer==2.2.0
audiofile==1.5.0
audmath==1.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
decorator==4.4.2
gTTS==2.5.4
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
lxml==5.3.0
moviepy==1.0.3
numpy==2.0.2
opencv-python==4.10.0.84
pillow==10.4.0
proglog==0.1.10
pycparser==2.22
pydub==0.25.1
python-dotenv==1.0.1
python-pptx==1.0.2
requests==2.32.3
scipy==1.13.1
soundfile==0.12.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
XlsxWriter==3.2.0
